Update weather.py - deprecated constants
TEMP_CELSIUS was used from metnowcast, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfTemperature.CELSIUS instead, please report it to the author of the 'metnowcast' custom integration
SPEED_METERS_PER_SECOND was used from metnowcast, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfSpeed.METERS_PER_SECOND instead, please report it to the author of the 'metnowcast' custom integration
LENGTH_MILLIMETERS was used from metnowcast, this is a deprecated constant which will be removed in HA Core 2025.1. Use UnitOfLength.MILLIMETERS instead, please report it to the author of the 'metnowcast' custom integration